Submission to Spring Battle of the Code Hackathon
I worked on this with my group to create an app for android which would
analyze receipts. The concept was to utilize the keywords retrieved from
the optical recognition to give the user an advertising experience
geared towards them. The incentive to utilize the app was to track money
spent on categories like gas, food, etc throughout the day/week.